basex_api-dev-sign/requirements.txt
redlock-py==1.0.8
socket.io-emitter==0.1.5.1
pymongo==4.1.1
celery==5.2.7
pydash==5.0.0
aiohttp==3.8.3
requests==2.28.2
marshmallow==3.16.0
Werkzeug==2.2.2
Flask==2.1.2
python-dotenv==0.16.0
motor==3.0.0
PyJWT==2.6.0
cryptography==38.0.3
bcrypt==4.0.1
web3==6.0.0
tweepy==4.14.0
python-slugify==8.0.1
slugify==0.0.1
Flask-Cors==4.0.0